deps: remove pandas, deprecret and chex from default deps
