+++ .github/workflows/code-quality.yaml
@@ -35,21 +35,21 @@ jobs:
         steps:
             - uses: actions/checkout@v4
             - uses: ./.github/actions/setup
-            - run: uv run --python 3.13 --with '.[test]' pytest tests/. -vv --durations=0
+            - run: uv run --python 3.13 --group test pytest tests/. -vv --durations=0
     tests_py3-12:
         runs-on: ubuntu-latest
         needs: [lock_file]
         steps:
             - uses: actions/checkout@v4
             - uses: ./.github/actions/setup
-            - run: uv run --python 3.12 --with '.[test]' pytest tests/. -vv --durations=0
+            - run: uv run --python 3.12 --group test pytest tests/. -vv --durations=0
     tests_py3-11:
         runs-on: ubuntu-latest
         needs: [lock_file]
         steps:
             - uses: actions/checkout@v4
             - uses: ./.github/actions/setup
-            - run: uv run --python 3.11 --with '.[test]' pytest tests/. -vv --durations=0
+            - run: uv run --python 3.11 --group test pytest tests/. -vv --durations=0
     build:
         runs-on: ubuntu-latest
         needs:
